Make Scope ID RNG no longer use SecureRandom
SecureRandom is a robust RNG for use in security-sensitive scenarios
(like generating cryptographic/private keys). For something like a
random Scope ID, it is overkill, as it's slower, and becomes a
potentially blocking call if the machine hasn't been on for long or have
enough entropy.

It's much better to just use a normal ThreadLocalRandom instance, that
will not experience contention issues even under high traffic.