load extensions from config endpoint (DAT-18666) (#6323)
Previously, the list of extensions that we wished to capture the versions of was hardcoded in the code. This means that if an extension is added, old versions of Liquibase would not be able to capture the version of that extension. This PR changes the behavior to store that list of extensions (and the names that the extensions have in their MANIFEST files) in the config endpoint, so it can be changed remotely.