Deployment ID enhancements and execution property (#5653)
This PR is created in response to challenges encountered in the creation development of #5598.

The purpose of this PR is to create a deployment ID that does have to be changed/generated multiple times and is available early on in the execution timeline.

This will allow the implementation of an execution property as setup in #5598 to be created for the deployment ID.

However, the deployment ID was generated later in the process to be a time-based number that was obtained after a database lock was obtained to guarantee uniqueness. This PR aims to change the deployment ID from a time-based number to a UUIDv4 to ensure uniqueness. The UUID (36 characters) does not fit in the existing databasechangelog.deployment_id column, so the column is updated in this PR to be a length of 36 (and is updated for both creation of the table and update of existing tables).

For a full discussion of why this came to be, the discussion on #5598 can be referenced.

---------

Co-authored-by: Jason Lyle <jason.lyle@oracle.com>
Co-authored-by: Daniel Mallorga <dmallorga@liquibase.com>