Implement logic to selectively include/exclude columns in generate-changelog "data" export (Issue #3371, #6310) (#6371)
* Implement logic to selectively include/exclude columns in generate-changelog "data" export

Both variants are supported: either includeObjects or excludeObjects.  The columns are taken from "column:PATTERN" or, if both "table:" and "column:" are missing, then the entire includeObjects / excludeObjects is used to match the columns to export.  If neither includeObjects nor excludeObjects are provided, all columns are considered, which is the default case.  This corresponds to todays situation.

This PR addresses issues #3310 and #6310 - ey, exaclt 3000 issues apart.