fix: Allow skipping a missing custom change
Prior to this a custom change with a class name that is not on the
classpath would always fail the entire changelog.

Now `changeSet.failOnError: false` will be honoured if the cause of the
failure is the failure to instantiate the change class.

In addition, a failing precondition for the changeset with
onFail="MARK_RAN" will be honoured if the class is not on the classpath.

Fixes https://github.com/liquibase/liquibase/issues/6520

Signed-off-by: Robert Elliot <rob@lidalia.org.uk>