remove recursive checking for multiple overrides of command steps (DAT-17162) (#6713)
Originally, I added this code in #6675 in order to support extensions overriding a pro command step which overrides an OSS command step. However, this pattern is not recommended. If two different extensions override the same command step and are both on the classpath at the same time, an exception is thrown. Thus, there isn't a useful way to use this feature, and as a result, I'm removing it so that we don't accidentally use it and cause a situation where multiple extensions conflict with each other.