Snapshot all configuration values for Oracle sequences (#6692)
In commit 88beffd sequence configuration values with Oracle defaults
have been converted to null to exclude them from the change log.
Unfortunatelly this broke the ability to create correct diffs in the
situation where the new state should be the default value (which is
currently represented as null and therefore not included in the change
log).

Restore capturing of all attributes like it is done for other databases
and elements.

Fixes #6632