make all command names public (#6932)
without this it is not possible to use the constants from the outside,
in which case one instead has to manually write the magic string, which
is error-prone.

other command names are already public and can already be used from the
outside.

this has been done using search & replace and should've solved it for
all existing commands.

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>