Allow skipping a missing custom change (Copy of 6627) (#7390)
* fix: Allow skipping a missing custom change

Prior to this a custom change with a class name that is not on the
classpath would always fail the entire changelog.

Now `changeSet.failOnError: false` will be honoured if the cause of the
failure is the failure to instantiate the change class.

In addition, a failing precondition for the changeset with
onFail="MARK_RAN" will be honoured if the class is not on the classpath.

Fixes https://github.com/liquibase/liquibase/issues/6520

Signed-off-by: Robert Elliot <rob@lidalia.org.uk>

* Log not loadable on custom change validation

Also fixes failing test by loading the change in generateChecksum

Signed-off-by: Robert Elliot <rob@lidalia.org.uk>

* Added additional test assertions to existent tests.

* Log whenever unable to load custom change

Signed-off-by: Robert Elliot <rob@lidalia.org.uk>

* - CustomChangeWrapper null check added.
- CustomChangeWrapper validation error/warning added to keep consistency with other scenarios instead of throwing an exceptionto keep consistency with other scenarios.
- Integration tests fixed.

* Re-apply changes of login warning messages instead of throwing an exception.

* Log message updated.

---------

Signed-off-by: Robert Elliot <rob@lidalia.org.uk>
Co-authored-by: Robert Elliot <rob@lidalia.org.uk>